add smart cache, comment in admin, fix question vote, saving comment

+++ app/Http/Controllers/QuestionVoteController.php
@@ -20,14 +20,20 @@ class QuestionVoteController extends Controller
 
             $model = QuestionVotes::query()
                 ->where('question_id', $question->id)
-                ->where('vote', $data['vote'])
+                ->where('user_id', auth()->id())
                 ->first();
             if (!$model) {
                 $status = QuestionVotes::create([
                     'question_id' => $question->id, 'user_id' => $data['user_id'], 'vote' => $data['vote']
                 ]);
             } else {
-                $status = $model->delete();
+                if ($model->vote === $data['vote']){
+                    $status = $model->delete();
+                } else {
+                    $status = $model->update([
+                        'vote' => $data['vote']
+                    ]);
+                }
             }
         } else {
             $status = false;
